requirements.txt
# 多Agent股票顾问系统

langchain>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0

# 数据源
baostock>=0.8.8
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0

# 新闻爬取
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 环境配置
python-dotenv>=1.0.0

# CLI
rich>=13.0.0
typer>=0.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
streamlit
